+++ GabrielaRamirez/EjerciciosHackerrank/src/basics/IfElse.java
@@ -1,15 +1,7 @@
-import java.io.*;
-import java.math.*;
-import java.security.*;
-import java.text.*;
-import java.util.*;
-import java.util.concurrent.*;
-import java.util.regex.*;
+package basics;
+import java.util.Scanner;
 
-public class Solution {
-
-
-
+public class IfElse {
     private static final Scanner scanner = new Scanner(System.in);
 
     public static void main(String[] args) {
@@ -17,7 +9,7 @@ public class Solution {
         scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");
 
         scanner.close();
-        
+
         if (N%2==0 & 2<=N & N<=5){
             System.out.println("Not Weird");
         }
@@ -32,8 +24,6 @@ public class Solution {
         }
         else if (N%2!=0){
             System.out.println("Weird");
-        }       
+        }
     }
-
-    
 }
