+++ GabrielaRamirez/EjerciciosHackerrank/src/basics/InheritanceI.java
@@ -1,0 +1,32 @@
+package basics;
+
+class Animal{
+    void walk()
+    {
+        System.out.println("I am walking");
+    }
+}
+class Bird extends Animal
+{
+    void fly()
+    {
+        System.out.println("I am flying");
+    }
+
+    void sing()
+    {
+        System.out.println("I am singing");
+    }
+}
+
+public class InheritanceI{
+
+    public static void main(String args[]){
+
+        Bird bird = new Bird();
+        bird.walk();
+        bird.fly();
+        bird.sing();
+
+    }
+}
