+++ GabrielaRamirez/EjerciciosHackerrank/src/basics/StdinI.java
@@ -1,0 +1,15 @@
+package basics;
+import java.util.Scanner;
+
+public class StdinI {
+    public static void main(String[] args) {
+        Scanner scan = new Scanner(System.in);
+        int a = scan.nextInt();
+        int b = scan.nextInt();
+        int c = scan.nextInt();
+
+        System.out.println(a);
+        System.out.println(b);
+        System.out.println(c);
+    }
+}
