+++ GabrielaRamirez/EjerciciosHackerrank/src/basics/StdinII.java
@@ -1,7 +1,7 @@
-import java.util.*;
+package basics;
+import java.util.Scanner;
 
-public class Solution {
-
+public class StdinII {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
         int a = scan.nextInt();
